\name{rstruct-package}
\alias{rstruct-package}
\alias{rstruct}
\docType{package}
\title{\packageTitle{rstruct}}
\description{\packageDescription{rstruct}}
\section{Package Content}{\packageIndices{rstruct}}
\author{\packageAuthor{rstruct}}
\section{Maintainer}{\packageMaintainer{rstruct}}
\keyword{package}
